Auto hide UI when site opened with parameters.
Users probably dont want to see the full UI if they followed a link with
vidoes in it, they just want to play them, so we do that automatically.

Since the hiding of the ui happens via javascript this would usually
result in the full ui including the help text being visibile for a short
time when the page has loaded but the javascript has not executed which
looks weird. To prevent this the user interface starts out hidden
(displaying just a blank page) and is enabled or disabled by
javascript.
This has the drawback that the site takes a little longer to display
when not opened with parameters.
In a perfect world there would be a different html document for video
links which has the ui hidden initially where as the standard index.html
would still show the ui initially.